requirements.txt
# ─── Core ───────────────────────────────────────────
psycopg2-binary==2.9.9
pandas==2.2.3
numpy==1.26.4
sqlalchemy==2.0.31

# ─── ML / LLM / Agent ──────────────────────────────
openai>=1.17.0           # Qwen compatible
langchain==0.2.2
tiktoken==0.7.0
langchain-openai>=0.1.3

# ─── Web / Dashboard ───────────────────────────────
gradio==4.44.1
plotly==5.22.0
cachetools==5.3.3

# ─── Misc ──────────────────────────────────────────
jinja2==3.1.4
python-dotenv==1.0.1
